Refactor system specs
This separates each file's concerns while ensuring we can navigate
between the two resources.

It also uses the terminology we agreed on in which we refer to
"delegation" for Setting and "participant" for Delegation. This way we
don't leak the DB design to the UI and make it simpler for the user.