Copy decidim-consultation's vote button states
Apart from copying the large if/else block from
`question/_vote_button.html.erb` some bits need to be changed from the
`current_user` to `delegation.granter`. And things turn out not the be
simple.

`#allowed_to?` doesn't make it possible to only pass in a different
user. We are forced to pass in the `chain` argument as well, in which we
just copying the result we got from `#permission_class_chain`. Switching
this method's signature to use keyboard arguments or a hash of options
would solve this issue.