Ensure we close and remove the current modal
This fixes the issue where right after voting, opening the
delegations modal again and the closing it, the former modal was there.

I'm assuming this is because without closing and removing it the fact
that has a listener attached, makes it stay in the DOM and so when we
render the new one that one is still there. Or something alike.

What I fail to understand is why does the reveal modal end up at the
bottom of the DOM even though we render it within the div
"question-<question_id>-vote-button", which holds the contents that we
re-render through update_vote_button.js.erb. Something that Foundation
does, I guess.