Override 'Supports' vote counter with liquidvoting API votingResults value.
Naively gets votingResults.inFavor value on every page refresh.

We can get new result whenever we create or delete a vote, so in
those cases, and perhaps also with delegations, we could use the
votingResults.inFavor value we returned from those mutations.

Currently updates `Supports` count correctly for voting, with or
without a delegation in support of the vote.