Introduce a liquidvoting state object
This is meant to provide controllers a way to request all the relevant LV proposals state at the beginning of a request.

I've used a read-only struct; the idea is to keep it a simple snapshot, and if ever a controller wants an up-to-date
copy (after voting, for example), they can re-request it. We could make it mutable but then someone is responsible
to keep it in sync with the LV api.

The expected lifetime is just the current web request.