Add fails to unused? VotesController
I think this was added, not a decidim override, and I think the functionality is now implemented in ProposalVotesController and
a couple of commands. I've bombed the actions so we can see if they get used; otherwise perhaps we remove the file.
There is some session and flash behavior we may want to keep.