Manage proposal_votes_count attribute (un)vote
This is to back out changes to completely abandon the proposals.proposal_votes_count attribute.

Instead, while we are still abandoning the ProposalVote model, we now "own" and manage the count attribute
by setting it with the current vote count from LV, whenever we vote or unvote. The best way to keep the rest
of Decidim working (things like sorting proposals on vote count) is to just keep the attribute up-to-date with
the actual vote count, per LV.

(will need to do same when creating/deleting delegations).