Enable VCR (WIP, broken)
I'm mostly following the example from decidim-elections >0.24 but this is not yet
generating a cassette as intended.

I've added a `VCR.use_cassette('create_vote')` ... decidim-elections doesn't need this, hopefully
we can fall back to just `describe ... :vcr` metadata approach.

Currently failing on the fetch_schema bit. We could copy the cassette from ruby-client, but it would
be better if we can generate-on-absence, so we have an easy way to refresh cassettes. I'm unsure
how to wrap that api call, maybe in a :before block.