Change api_state to carry delegate id not email
The trigger was to stop carrying the delegate's email as a form params (needed when the user undelegates).

The consequence is we now use delegate id rather than email in api_state, and to drive all delegate-related rendering.

We are still using delegate_email in the Liquidvoting interface, as well as the api's interface.

Given that our api allows voting and delegating by either emails or ids (uuids I think), and given that
email is a sensitive field, I think delegate (Decidim::User) is a better argument for the Liquidvoting interface
than delegate_email.

And separately we might move to only using ids rather than emails in the api's interface.