Make #user_proposal_state proposal_url optional
In the case of ProposalsController#index, we don't have a specific proposal in mind
(rather we have a bunch of proposal cards).

But Liquidvoting.user_proposal_state needed a proposal_url to get api state.

This commit makes proposal_url optional, returning a non-proposal-specific state object,
which makes the views Happy Again.