Fix the proposal cell footer
We're overriding the cell footer, which needs api_state before rendering the vote_button.

This should be done by a controller-like thing, or a presenter, but I don't know how to do that with cells yet.

Instead, I've just hacked in the api_state call, and pass it as a local to the vote_button partial