Add js check of delected delegate to UI related ajax
This ensures the js check for no delegate selected, when click
submit on delegation form, is loaded when the ajax in
update_buttons_and_counters.js.erb is run.

Ideally, we should not copy the code block into this file, but
should call the code as a function from another file.